+++ sam/app/annotations/__init__.py
@@ -1,0 +1,3 @@
+"""Segment Anything API for Cytomine."""
+
+__version__ = "0.1.0"
